web_app_gcp/app/v4/requirements.txt
fastapi[standard]>=0.112.0
PyMySQL==1.1.1
google-cloud-storage==2.17.0
google-cloud-pubsub==2.21.4
redis==5.0.7
